fix(metrics): declare tool widgets in the central posthogAiToolRenderers manifest
Master moved tool-widget registration out of bootApp into a central manifest (#97264), which conflicted with this branch's bootApp registration. The metrics entries now live in products/metrics/frontend/posthogAiToolRenderers.tsx, listed in frontend/src/posthogAiToolRenderers.ts, matching the logs, error-tracking and cdp products. The output parser reads structured tool results through getToolOutputRecord, replacing the removed parseToolOutputRecord (#97494).

Generated-By: PostHog Desktop
Task-Id: 1b494f2c-d6ab-4a9a-9297-4e0f2c12597a
