+++ frontend/src/posthogAiToolRenderers.ts
@@ -1,6 +1,7 @@
 import { posthogAiToolRenderers as cdpToolRenderers } from 'products/cdp/frontend/posthogAiToolRenderers'
 import { posthogAiToolRenderers as errorTrackingToolRenderers } from 'products/error_tracking/frontend/posthogAiToolRenderers'
 import { posthogAiToolRenderers as logsToolRenderers } from 'products/logs/frontend/posthogAiToolRenderers'
+import { posthogAiToolRenderers as metricsToolRenderers } from 'products/metrics/frontend/posthogAiToolRenderers'
 import type { ToolRegistryEntry } from 'products/posthog_ai/frontend/api/tools'
 import { posthogAiToolRenderers as dataToolRenderers } from 'products/posthog_ai/frontend/posthogAiToolRenderers'
 
@@ -9,4 +10,5 @@ export const posthogAiToolRenderers: ToolRegistryEntry[] = [
     ...cdpToolRenderers,
     ...errorTrackingToolRenderers,
     ...logsToolRenderers,
+    ...metricsToolRenderers,
 ]
