+++ products/metrics/frontend/posthogAiToolRenderers.test.ts
@@ -3,13 +3,11 @@ import path from 'path'
 
 import { toolRegistry } from 'products/posthog_ai/frontend/api/tools'
 
-import { registerMetricsToolRenderers } from './registerMetricsToolRenderers'
-
 // Every metrics tool the MCP codegen emits, so a tool added to the catalog without a registry entry
 // fails here instead of shipping the generic wrench card. Read off the generated file rather than
 // imported, to keep the services/mcp graph out of a frontend test.
 function generatedMetricsToolNames(): string[] {
-    const generated = path.resolve(__dirname, '../../../../services/mcp/src/tools/generated/metrics.ts')
+    const generated = path.resolve(__dirname, '../../../services/mcp/src/tools/generated/metrics.ts')
     const source = fs.readFileSync(generated, 'utf-8')
     const names = [...source.matchAll(/name: '([a-z0-9-]+)'/g)].map((m) => m[1])
     if (names.length === 0) {
@@ -18,11 +16,7 @@ function generatedMetricsToolNames(): string[] {
     return [...new Set(names)].sort()
 }
 
-describe('registerMetricsToolRenderers', () => {
-    beforeAll(() => {
-        registerMetricsToolRenderers()
-    })
-
+describe('posthogAiToolRenderers', () => {
     it.each(generatedMetricsToolNames())('claims %s in the shared registry', (key) => {
         expect(toolRegistry.lookup(key)).not.toBeNull()
     })
